refactor: update import paths for hooks in various components and remove unused fetch hooks
