+++ src/features/notes/hooks/useUploadNote.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useToast } from "./use-toast";
+import { useToast } from "../../../hooks/use-toast";
 import axiosClient from "@/lib/axiosClient";
 const uploadNote = async (data) => {
   console.log(data);
