+++ src/features/subtasks/hooks/useSubTaskIsComplete.js
@@ -2,7 +2,7 @@
 
 import axios from "axios";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { useToast } from "./use-toast";
+import { useToast } from "../../../hooks/use-toast";
 import axiosClient from "@/lib/axiosClient";
 
 const useSubTaskIsComplete = (getToken) => {
