+++ src/features/tasks/TaskModal.js
@@ -19,8 +19,8 @@ import { Checkbox } from "@/components/ui/checkbox";
 import { useAuth } from "@clerk/nextjs";
 import { useQueryClient } from "@tanstack/react-query";
 import useTaskPositionUpdate from "@/features/tasks/hooks/useTaskPositionUpdate";
-import { singleSubTask } from "@/hooks/useFetchSingleSubTask";
-import { singleNote } from "@/hooks/useFetchSingleNote";
+import { singleSubTask } from "@/features/subtasks/hooks/useFetchSingleSubTask";
+import { singleNote } from "@/features/notes/hooks/useFetchSingleNote";
 import { TaskCollapsibleButton } from "./TaskCollapsibleButton";
 import TaskDialogCard from "./TaskDialogCard";
 import useIsComplete from "./hooks/useIsComplete";
